+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `all` command added - Creates changelog for markdown, rst and json at the same time.
+
     
 ## [1.0.6] - 2018-03-01