+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - `write_*()` -> `write()`. `write_*` will be deprecated in version 1.2
+- `requests.py` -> `request.py`
     
 ## [1.0.6] - 2018-03-01