+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `write_*()` -> `write()`. `write_*` will be deprecated in version 1.2
 - `requests.py` -> `request.py`
+
+### Removed
+
+- `FormatRstBase` removed and merged with `FormatBase`.
     
 ## [1.0.6] - 2018-03-01