+++ CHANGELOG.md
@@ -13,11 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `deprecate` added in `utils.py`.
 - `check_version` added in `utils.py` and added as thread call to CLI
 
-
 ### Changed
 
 - `write_*()` -> `write()`. `write_*` will be deprecated in version 1.2
 - `requests.py` -> `request.py`
+- `README.rst` -> `README.md`
 
 ### Removed